pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ummfiltered"
version = "0.1.0"
description = "Remove filler words from talking-head videos"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "rich>=13.0.0",
    "scikit-image>=0.22.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
cloud = ["deepgram-sdk>=3.0.0", "openai>=1.0.0"]
gpu = ["torch>=2.0.0"]
dev = ["pytest>=7.0.0", "pytest-mock>=3.0.0"]

[project.scripts]
ummfiltered = "ummfiltered.cli:main"
